backend-djangorest/djangorest/src/unversioned-requirements.txt
django
# Rest framework
djangorestframework
# Environment variables library
django-environ
# More static types and type inference
django-stubs
# Multiple configurations inside one setting
django-configurations
# Filtering by url
django-filter
# Cors
django-cors-headers
# Swagger
drf-yasg
# Jwt Authentication
djangorestframework-simplejwt
# Jwt encription
PyJWT
cryptography
# Image library
Pillow
# Celery driver
celery
# Celery task result models
django-celery-results
# Postgres driver
psycopg2-binary
# Testing coverage
coverage
# Celery task result models
django-celery-results
# Celery periodic task models
django-celery-beat
# Redis driver
redis
# Concurrent networking library for Celery
cod
# Web server
gunicorn
# Cryptographic library
pycryptodome